requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0
ta-lib>=0.4.0
yfinance>=0.2.0

# AI/ML Framework
tensorflow>=2.13.0
torch>=2.0.0
scikit-learn>=1.3.0
xgboost>=1.7.0
stable-baselines3>=2.0.0
finrl>=0.3.0

# Trading APIs
ccxt>=4.0.0
alpaca-trade-api>=3.0.0
python-binance>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0
alembic>=1.11.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.22.0
websockets>=11.0.0

# Task Scheduling
celery>=5.3.0
flower>=2.0.0
apscheduler>=3.10.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0.0

# Logging & Monitoring
loguru>=0.7.0
prometheus-client>=0.17.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
faker>=19.0.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.0
pytz>=2023.3
arrow>=1.2.0
rich>=13.0.0
click>=8.1.0

# Technical Analysis
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data Validation
marshmallow>=3.20.0
cerberus>=1.3.0

# Async Support
aiohttp>=3.8.0
asyncio-mqtt>=0.14.0
aiofiles>=23.0.0

# Security
cryptography>=41.0.0
bcrypt>=4.0.0
python-jose>=3.3.0

# Monitoring
sentry-sdk>=1.30.0